requirements.txt
streamlit>=1.31.0
PyPDF2>=3.0.1

langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.4

sentence-transformers>=2.6.1
faiss-cpu>=1.7.4

numpy>=1.24.0



#langchain-huggingface-hub
#sentence-transformers

